+++ include/CollisionManager.hpp
@@ -15,8 +15,14 @@
 #include "Events/ActorCreatedEvent.hpp"
 #include "Events/ActorDestroyedEvent.hpp"
 
+using std::cerr;
+//int pair (used to assocaite a row and col)
+typedef pair<int,int> intPair;
+//int pair (used to assocaite a x and y)
+typedef pair<float,float> floatPair;
 class CollisionManager{
 private:
+
   //Store the textLoader to make requests for strings and constants
   shared_ptr<TextLoader> textLoader;
   //Store the Event Manger for the game that receives and distributes irregular
@@ -46,6 +52,11 @@ private:
   //the number of cols
   int cols;
 
+  //rectangle shape to act in place of the tiles to check if a circle intersects
+  sf::RectangleShape tile;
+
+  int nummade = 0;
+  int numdestroyed = 0;
 public:
 
   CollisionManager(shared_ptr<TextLoader> textLoader, shared_ptr<EventManager> eventManager,
@@ -68,6 +79,11 @@ public:
 
   void addNewTower(shared_ptr<TowerInterface> tower);
   void removeOldTower(shared_ptr<TowerInterface> tower);
+  void handleTowerRadiusUpgrade(shared_ptr<TowerInterface> towerUpgraded);
+  void addTowerByRadius(shared_ptr<TowerInterface> tower);
+  vector<intPair> minAndMaxRowCol(shared_ptr<TowerInterface> tower);
+  bool tileInRadius(shared_ptr<TowerInterface> tower);
+  vector<floatPair> getFourCornersOfTileAndMidpoints();
 
   void checkForCollisions();
   void updateAllUnitPositions();