+++ src/MeleeUnit.cpp
@@ -92,6 +92,10 @@ HitpointBar MeleeUnit::getHpBar() {
 void MeleeUnit::move(float deltaS){
   this->x+=(speed*cos(direction))*ActorInterface::getXScale();
   this->y+=(speed*sin(direction)*-1)*ActorInterface::getYScale();
+  if(isnan(x) || isnan(y)){
+    cout << "position after " << x << " " << y << endl;
+    cout << "direction was " << direction << endl;
+  }
 }
 
 bool MeleeUnit::isCollision(sf::FloatRect colliding_with){