+++ src/WaveManager.cpp
@@ -148,7 +148,7 @@ void WaveManager::createNextWave() {
   distancesFromEntryPositions=getNormalizedDistanceMap(distancesFromEntryPositions);
 
   for(auto it=distancesFromEntryPositions.begin();it!=distancesFromEntryPositions.end();it++){
-    cout << "map dist normalized " << it->first << endl;
+    //cout << "map dist normalized " << it->first << endl;
   }
 
   double range=(--distancesFromEntryPositions.end())->first;
@@ -192,7 +192,7 @@ void WaveManager::createNextWave() {
     }else if(spawn_distance>range){
       spawn_distance=range-.1;
     }
-    cout << "spawn dist: " << spawn_distance << endl;
+    //cout << "spawn dist: " << spawn_distance << endl;
 
     int roundedKey;