+++ Changelog.md
@@ -24,6 +24,8 @@
 - yes, I did all this... it took many hours
 
 ###Changed
+- going below 0 hitpoints ends the game [@andrewkatson][https://github.com/andrewkatson]
+- towers go into the right row, col in the collision manager
 - Improved unit animations [@jeremyelkayam][https://github.com/jeremyelkayam]
 - OptionsMenuScreen now sends events to player, boardmanager, and wavemanager [kp-hb][https://github.com/kp-hb]
 - Player has a handleoptionselectedevent method